FrameMixin moved to guicore/mapdisp
